+++ packages/server/src/http/index.ts
@@ -2,7 +2,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { assertNotBrowser } from '../assertNotBrowser';
 
-export * from './requestHandler';
+export * from './resolveHttpResponse';
 export * from './ResponseMeta';
 
 assertNotBrowser();
